Move RedisKey type
